Separate out modules from login
Login should be entirely independent module and only after login it
should route to main app
